Relocated files, package structure for rest-apiv2 Removed build files
