+++ enme-mvc/rest-api/api-v2/src/main/java/org/encuestame/rest/api/v2/package-info.java
@@ -2,4 +2,4 @@
  * Contains class and implement the Mail Delivery Service.
  *
  */
-package org.encuestame.mvc.controller.json.v2;
+package org.encuestame.rest.api.v2;
