--- enme-utils/build.gradle
@@ -1,22 +1,0 @@
-
-description = 'Encuestame Utilities'
-dependencies {
-    compile group: 'junit', name: 'junit', version:'4.12'
-    compile group: 'commons-collections', name: 'commons-collections', version:'3.2.1'
-    compile group: 'commons-lang', name: 'commons-lang', version:'2.4'
-    compile group: 'joda-time', name: 'joda-time', version:'1.6.2'
-    compile group: 'commons-logging', name: 'commons-logging', version:'1.2'
-    compile group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version:'2.6.0'
-    compile group: 'commons-io', name: 'commons-io', version:'2.0.1'
-    compile(group: 'javax.servlet', name: 'javax.servlet-api', version:'3.0.1') {
-       /* This dependency was originally in the Maven provided scope, but the project was not of type war.
-       This behavior is not yet supported by Gradle, so this dependency has been converted to a compile dependency.
-       Please review and delete this closure when resolved. */
-    }
-}
-
-task packageTests(type: Jar) {
-  from sourceSets.test.output
-  classifier = 'tests'
-}
-artifacts.archives packageTests
