--- enme-utils/src/site/site.xml
@@ -1,91 +1,0 @@
-<?xml version="1.0" encoding="ISO-8859-1"?>
-<project name="encuestame">
-    <bannerLeft>
-        <name>Encuestame</name>
-        <src>http://www.encuestame.org/images/encuestame_header.png</src>
-        <href>http://www.encuestame.org/</href>
-    </bannerLeft>
-    <bannerRight>
-         <src>http://www.encuestame.org/images/twitter.jpg</src>
-        <href>http://www.twitter.com/encuestame</href>
-    </bannerRight>
-
-    <skin>
-        <groupId>org.apache.maven.skins</groupId>
-        <artifactId>maven-default-skin</artifactId>
-        <version>1.0</version>
-    </skin>
-
-
-    <publishDate format="dd MMM yyyy" position="left" />
-    <version position="left" />
-
-    <body>
-
-        <head>
-            <!-- Google Analytics - Start -->
-            <script type="text/javascript">
-
-                var _gaq = _gaq || [];
-                _gaq.push(['_setAccount', 'UA-2527438-15']);
-                _gaq.push(['_trackPageview']);
-
-                (function() {
-                var ga = document.createElement('script'); ga.type = 'text/javascript';
-                ga.async = true;
-                ga.src = ('https:' == document.location.protocol ? 'https://ssl' :
-                'http://www') + '.google-analytics.com/ga.js';
-                var s = document.getElementsByTagName('script')[0];
-                s.parentNode.insertBefore(ga, s);
-                })();
-
-        </script>
-        <!-- Google Analytics - End -->
-
-        <!-- SEO -->
-        <title>Encuestame Java Open Source Survey System, Twitter, Poll, Surveys.</title>
-        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-        <meta name="description" content="Open Source Java Survey System, Twitter, Poll, Surveys." />
-        <meta name="keywords" content="survey, java, open source, poll, twitter, spring framework" />
-        <meta name="google-site-verification" content="Cp-ow0L1hAqhoCF_7-og3NgKqUk01S-Ln_JQkZV5Bx0" />
-        <meta http-equiv="Expires" content="never"/>
-        <meta name="Revisit-After" content="2"/>
-        <meta name="distribution" content="Global"/>
-        <meta name="Robots" content="INDEX,FOLLOW"/>
-        <meta name="author" content="encuestame.org"/>
-        <!-- SEO - End -->
-
-        </head>
-        <!-- more stuff -->
-
-        <!--
-        <breadcrumbs>
-            <item href="" name="" />
-            <item name="" />
-        </breadcrumbs>
-        -->
-
-        <links>
-            <item name="Main" href="../index.html" />
-            <item name="Download" href="../download.html" />
-            <item name="Features" href="../features.html" />
-            <item name="Users Forum" href="http://forum.encuestame.org" />
-            <item name="Sourceforge.net Project Page" href="http://sourceforge.net/projects/encuestame/" />
-        </links>
-
-        <menu name="Encuestame Project">
-            <item name="Main" href="index.html" />
-            <item name="License" href="license.html" />
-        </menu>
-
-        <menu ref="reports" />
-
-        <menu ref="modules" />
-
-    </body>
-
-    <poweredBy>
-        <logo name="Hosted on SourceForge.net" href="http://sourceforge.net/projects/encuestame/"
-            img="http://sflogo.sourceforge.net/sflogo.php?group_id=183612&amp;type=1" />
-    </poweredBy>
-</project>
